Stop saving audit model without any info
Things can get messy if the audit model is created (to get a pk for the
vendor_tx_code) but then the process crashes. This change ensures that
audit models are only created once we have something interesting to
store in them.

It requires some fiddling with the mocks to ensure the database isn't
used.